requirements.txt
# Core
streamlit
python-dotenv

# AI Models / API
ultralytics
google-generativeai

# Image Processing
pillow
opencv-python-headless

# Utilities
requests
reportlab
